# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quantum FX Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /team/earnings-overview:
    get:
      operationId: getTeamEarningsOverview
      tags: [team]
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [this_month, last_month, all_time]
      responses:
        "200":
          description: Team earnings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamEarnings"

  # Profile
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile/change-password:
    post:
      operationId: changePassword
      tags: [profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profile/2fa:
    get:
      operationId: get2FA
      tags: [profile]
      responses:
        "200":
          description: 2FA status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwoFAStatus"
    post:
      operationId: toggle2FA
      tags: [profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFAToggleInput"
      responses:
        "200":
          description: 2FA toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwoFAStatus"
  /profile/kyc:
    get:
      operationId: getKYC
      tags: [profile]
      responses:
        "200":
          description: KYC status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KYCStatus"
    post:
      operationId: submitKYC
      tags: [profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KYCInput"
      responses:
        "200":
          description: KYC submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KYCStatus"
  /profile/notification-settings:
    get:
      operationId: getNotificationSettings
      tags: [profile]
      responses:
        "200":
          description: Notification settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
    patch:
      operationId: updateNotificationSettings
      tags: [profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
...
  /profile/sessions/{id}:
    delete:
      operationId: revokeSession
      tags: [profile]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Notifications
...
  /support/tickets:
    get:
      operationId: listSupportTickets
      tags: [support]
      responses:
        "200":
          description: Support tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
    post:
      operationId: createSupportTicket
      tags: [support]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketInput"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Trade
...
    ChartPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: number
      required: [date, value]
...
    TeamEarnings:
      type: object
      properties:
        period:
          type: string
        totalEarnings:
          type: number
        chart:
          type: array
          items:
            $ref: "#/components/schemas/ChartPoint"
      required: [period, totalEarnings, chart]

    # Profile
    Profile:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        kycStatus:
          type: string
        twoFAEnabled:
          type: boolean
        createdAt:
          type: string
      required: [id, fullName, email, kycStatus, twoFAEnabled, createdAt]
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        country:
          type: string
    ChangePasswordInput:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]
    TwoFAStatus:
      type: object
      properties:
        enabled:
          type: boolean
        qrCode:
          type: ["string", "null"]
      required: [enabled]
    TwoFAToggleInput:
      type: object
      properties:
        enable:
          type: boolean
        code:
          type: ["string", "null"]
      required: [enable]
    KYCStatus:
      type: object
      properties:
        status:
          type: string
        submittedAt:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
      required: [status]
    KYCInput:
      type: object
      properties:
        documentType:
          type: string
        documentFrontUrl:
          type: string
        selfieUrl:
          type: string
        proofOfAddressUrl:
          type: ["string", "null"]
      required: [documentType, documentFrontUrl, selfieUrl]
    NotificationSettings:
      type: object
      properties:
        emailNotifications:
          type: boolean
        botAlerts:
          type: boolean
        depositWithdrawal:
          type: boolean
        promotions:
          type: boolean
      required: [emailNotifications, botAlerts, depositWithdrawal, promotions]
    NotificationSettingsUpdate:
      type: object
      properties:
        emailNotifications:
          type: boolean
        botAlerts:
          type: boolean
        depositWithdrawal:
          type: boolean
        promotions:
          type: boolean
...
    SupportTicket:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        message:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, subject, message, status, createdAt, updatedAt]
    SupportTicketInput:
      type: object
      properties:
        subject:
          type: string
        message:
          type: string
        category:
          type: string
      required: [subject, message, category]
...